Rename: 기존 카테고리 클래스명 변경
