+++ src/main/java/JesusDeciples/RankingQuiz/api/webSocket/VocaQuizDataCenterMediator.java
@@ -2,38 +2,36 @@ package JesusDeciples.RankingQuiz.api.webSocket;
 
 import JesusDeciples.RankingQuiz.api.dto.QuizDto;
 import JesusDeciples.RankingQuiz.api.dto.response.QuizResultDto;
-import JesusDeciples.RankingQuiz.api.service.quizDataCenter.VocaQuizDataCenter;
 import JesusDeciples.RankingQuiz.api.service.quizDataCenter.state.DataCenterState;
+import JesusDeciples.RankingQuiz.api.service.quizDataCenter.voca.VocaQuizDataCenter;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
 
-@Slf4j
 @Component
 @RequiredArgsConstructor
 public class VocaQuizDataCenterMediator {
-    private final VocaQuizDataCenter vocaQuizDataCenter;
+    private final VocaQuizDataCenter quizDataCenter;
 
     public DataCenterState getQuizDataCenterState() {
-        return vocaQuizDataCenter.getPresentState();
+        return quizDataCenter.getPresentState();
     }
 
     public void updateDataCenterStateAndAction(DataCenterState dataCenterState) {
-        vocaQuizDataCenter.setPresentState(dataCenterState);
-        vocaQuizDataCenter.handle();
+        quizDataCenter.setPresentState(dataCenterState);
+        quizDataCenter.handle();
     }
 
     public Map<String, QuizResultDto> getQuizResults() {
-        return vocaQuizDataCenter.getResults();
+        return quizDataCenter.getResults();
     }
 
     public QuizDto getPresentQuizDto() {
-        return vocaQuizDataCenter.getPresentQuizDto();
+        return quizDataCenter.getPresentQuizDto();
     }
 
     public String getQuizWinnerName() {
-        return vocaQuizDataCenter.getWinnerName();
+        return quizDataCenter.getWinnerName();
     }
 }
